Refactor utils structure
