+++ src/components/DropDown/DropDown.js
@@ -4,8 +4,9 @@ import styled from 'styled-components'
 import { Motion, spring } from 'react-motion'
 import ClickOutHandler from 'react-onclickout'
 import theme from '../../theme'
-import { springConf, unselectable } from '../../shared-styles'
-import { lerp } from '../../math-utils'
+import { spring as springConf } from '../../utils/styles/spring'
+import { unselectable } from '../../utils/styles'
+import { lerp } from '../../utils/math'
 import getPublicUrl, { styledPublicUrl as asset } from '../../public-url'
 import DropDownItem from './DropDownItem'
 import arrow from './assets/arrow-down.svg'
