+++ src/components/Header/Header.js
@@ -3,7 +3,7 @@ import type { MenuItemRenderLink } from './MenuItem'
 import React from 'react'
 import styled from 'styled-components'
 import theme from '../../theme'
-import { breakpoint, BreakPoint } from '../../shared-styles'
+import { breakpoint, BreakPoint } from '../../utils/styles/breakpoint'
 import Button from '../Button/Button'
 import Text from '../Text/Text'
 import MenuItem from './MenuItem'
