+++ src/components/Header/MenuPanel.js
@@ -4,7 +4,8 @@ import React from 'react'
 import styled from 'styled-components'
 import { Motion, spring } from 'react-motion'
 import ClickOutHandler from 'react-onclickout'
-import { spring as springConf, unselectable } from '../../shared-styles'
+import { spring as springConf } from '../../utils/styles/spring'
+import { unselectable } from '../../utils/styles'
 import getPublicUrl, { prefixUrl } from '../../public-url'
 import theme from '../../theme'
