+++ src/components/SidePanel/SidePanel.js
@@ -4,8 +4,9 @@ import React from 'react'
 import styled from 'styled-components'
 import { Motion, spring } from 'react-motion'
 import Text from '../Text/Text'
-import { lerp } from '../../math-utils'
-import { springConf, unselectable } from '../../shared-styles'
+import { lerp } from '../../utils/math'
+import { spring as springConf } from '../../utils/styles/spring'
+import { unselectable } from '../../utils/styles'
 import getPublicUrl, { prefixUrl } from '../../public-url'
 
 import close from './assets/close.svg'
