+++ src/utils/styles/index.js
@@ -1,0 +1,8 @@
+/* @flow */
+export const unselectable = () => `
+  -webkit-touch-callout: none;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+`
